Create this file
The Gram matrix is calculated based on the feature maps extracted from the VGG network. Each feature map has a shape of (channel, height, width), and to calculate the Gram matrix, we need to reshape it into a matrix of size (channel, height * width) and then take the matrix multiplication with its transpose.

If the input image size is not fixed, then the size of the feature maps extracted from the VGG network will also be different for each input image. As a result, the size of the Gram matrix will also vary, and this will cause the Style Loss to become inconsistent across different input images.